// library-app/src/models/Page.ts
export type PageInfo = {
    totalCount: number;     // Total number of items
    currentPage: number;    // The current page number
    totalPages: number;     // Total number of pages
    limit: number;          // Number of items per page
    pageCount: number;      // Number of items on the current page
  };
  